Fix Ticket #549: Target highlights not always working.
This module had a number of logic problems with it.  It had a mix of
multiple design patterns that we use in PB4 modules.  Some parts of it
assumed that the highlight was only on the frame when it was shown and
some parts tried to hide and show the frame.

Includes a slight optomization, as the color for the highlight is now
only set on frame creation and the configuration code nows finds all the
highlight frames and changes their colors when the config changes.

The standard method of operation now for this module is to leave the
Highlight frame in existance unless the module is disabled for the
layout or the frame is entirely torn down.  The frame will be made on
demand when it is first needed.