Fix the bug that causes the auto-hide CastBar to ocassionally not hide.
There was a rare race condition between GetColor and FixCastData.  As a
CastBar is fading out we bounce back and forth between these two
functions.  When the calculation of the alpha gets to or below zero then
the bar should be hidden.  GetColor would remove the cast_data and stop
the timer frame but not clear the frames.  As a result the CastBar
would be stuck.  Really, the GetColor function should not be deleting
entries in cast_data, that should always happen in FixCastData.