Make sure sizing is applied to GroupHeader's even when hidden.
It's impossible to anchor to the group headers before they're shown
because they have 0 sized height and width.

This changes ProxySetAttribute() to actually check if the frame is
visible before returning that an update actually happened.

It also changes Update to check if the frame is shown and if not shows
it before doing our noop proxy attribute to force an update, then hides
it again after.