Fix an error caused by recycling a BetterStatusBar with animations on.
Prior to this the animation flags were turned off when the control was
retrieved.  Which led to a bit of a chicken/egg syndrome.  The value
gets updated during the retrieval process and it thinks the animations
are turned on so it expects the animation controls to exist, which they
don't since they're recycled separately.

Now the flags are cleared when the animation objects are recycled.