Add support for the built in incoming heals data in 4.0.x.
This is implemented as a separate module so that LibHealComm will not be
loaded on 4.0.x.  Unless LibHealComm is installed as a separate addon or
embedded in another addon it will not be loaded on a 4.0.x client.

Both modules have the same loaded name in PitBull4, so they share the
same config.