Fix problems with click through frames.
- If you were using Clique singleton click through frames could become
clickable in combat.  The fix requires a version newer than Clique:
v40000-1.2.7 (which is the current release version at this time,
so in order to get the fix you need the alpha build on wowace.com).

- When creating frames in combat for groups with click through turned on
the frames would be clickable until the next time you left combat.
This was a limiation of WoW which has been resolved with 4.0.1.