When changing Unit group on Groups change is_wacky if necessary.
This should fix groups not working properly if you say duplicate the
Party frame and then set it to Party targets (a set of units we don't
receive events for).  Or vice versa (should prevent weird flickering)
when you take a Group that believes it is wacky and convert it to a
non-wacky set of units.  The issue was only temporary until you
reloaded but it did happen.

Yes my friends, PitBull4 bugs have gotten this edge casey.